backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4

# Caching & Queue
redis==5.0.1
celery==5.3.6
flower==2.0.1

# LLM Providers
openai==1.10.0
httpx==0.26.0

# Settings & Config
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Metrics
prometheus-client==0.19.0
python-json-logger==2.0.7

# WebSocket
websockets==12.0

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# RAG Document Processing
tiktoken==0.5.2
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1
chardet==5.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.1.14
